% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_as_date.R
\name{fast_as_date}
\alias{fast_as_date}
\title{Fast change to date class}
\usage{
fast_as_date(x, format = NULL)
}
\arguments{
\item{x}{a vector to be converted into date.}

\item{format}{the format of the dates. The default format is "2020-12-30", different format can be used but should be explicitly defined. For other formats of dates look at: \url{https://www.stat.berkeley.edu/~s133/dates.html}}
}
\value{
A vector of Dates in format of \code{YYYY-MM-DD}
}
\description{
This function is a wrapper function of the \code{lubridate::fast_strptime} function. It coverts a character or factor vector into a vector of dates.
}
\examples{
fast_as_date(c("2020-10-15","2020-12-03"))
fast_as_date(c("10/15/20","12/03/20"),format='\%m/\%d/\%y')
fast_as_date(c("October 15, 2020","December 03, 2020"),format='\%B \%d, \%Y')
}
\seealso{
see \code{\link[lubridate:parse_date_time]{lubridate::fast_strptime}} help for more information
}
