% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Find the statistical mode}
\usage{
getmode(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds. The default is FALSE.}
}
\value{
The value that appeared most in a set of data values
}
\description{
This function finds the statistical mode (the value that appears most often). In case of two modes, the first value will be return.
}
\examples{
getmode(c(1,1,2,3,NA,NA,NA))
getmode(c(1,1,2,3,NA,NA,NA),na.rm=TRUE)
getmode(c("A","B","C","A",NA,NA,NA),na.rm=TRUE)
}
\references{
Based on \url{https://stackoverflow.com/a/8189441}
}
