% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_unique.R
\name{n_unique}
\alias{n_unique}
\title{Count the number of unique values}
\usage{
n_unique(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of dates/factor levels/number}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds. The default is FALSE.}
}
\value{
The number of unique values in the vector
}
\description{
This function counts the number of unique values.
}
\examples{
n_unique(c(1,1,2,3,NA,NA,NA))
n_unique(c(1,1,2,3,NA,NA,NA),na.rm=TRUE)
n_unique(c("A","B","C","A",NA,NA,NA))
n_unique(c("A","B","C","A",NA,NA,NA),na.rm=TRUE)
}
