% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orar_theme.R
\name{orar_theme}
\alias{orar_theme}
\title{Customize Theme for ggplot2}
\usage{
orar_theme(
  legend_position = "bottom",
  legend_bg = "transparent",
  legend_direction = "horizontal",
  grid = "all",
  plot_ratio = 1.5,
  panel_aspect_ratio = NULL,
  distance_from_edge = 0.05,
  base_family = "sans",
  base_size = 20,
  ...
)
}
\arguments{
\item{legend_position}{A string specifying the position of the legend. Can be one of "bottom",
"top", "left", "right", or corner positions like "topleft", "topright", "bottomleft", "bottomright".}

\item{legend_bg}{A string specifying the background color of the legend box. "transparent" is the default.}

\item{legend_direction}{A string indicating the orientation of the legend. Can be "horizontal" or "vertical".}

\item{grid}{A string that controls the appearance of the grid in the plot. It can be "all", "none", "x", or "y".}

\item{plot_ratio}{Numeric. A ratio for the plot dimensions.}

\item{panel_aspect_ratio}{Numeric. An optional parameter for setting the aspect ratio of the plot panel.}

\item{distance_from_edge}{Numeric. A distance measure from the edge of the plot, used when legend is placed at corners.}

\item{base_family}{A string specifying the font family for the plot text.}

\item{base_size}{Numeric. Base font size for the plot text.}

\item{...}{Additional arguments passed to \code{theme_classic()}.}
}
\description{
This function \code{orar_theme} allows for advanced customization of ggplot2 plot themes,
including options for legend position, background, direction, and more.
}
\details{
The function \code{orar_theme} enhances the visual appearance of ggplot2 plots by providing
extensive customization options. It is particularly useful for adjusting legend properties
and grid settings. When specifying corner positions for the legend, the function automatically
adjusts its alignment and position relative to the plot edges.
}
\examples{
# Example usage of orar_theme
library(ggplot2)
p <- ggplot(mpg, aes(displ, hwy, colour = class)) + geom_point()
p + orar_theme(legend_position = "topright",
               legend_bg = "white",
               legend_direction = "vertical",
               grid = "all",
               base_family = "sans",
               base_size = 12)

# Using orar_theme with corner legend position
p + orar_theme(legend_position = "bottomright",
               legend_bg = "lightblue",
               grid = "y")

}
