% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installer.R
\name{package_installer}
\alias{package_installer}
\title{Fast and compact way to install/load libraries}
\usage{
package_installer(packages)
}
\arguments{
\item{packages}{a vector of packages names.}
}
\value{
The value that appeared most in a set of data values
}
\description{
This function take a vector of packages names as input and: install any package that is currently not installed.
Then, it loads all the packages.
}
\details{
The function install packages from CRAN repository. Other repositories including Github does not supported.
}
\examples{
package_installer(c("ggplot2","data.table"))
}
\references{
Based on \url{https://statsandr.com/blog/an-efficient-way-to-install-and-load-r-packages/}.
}
