% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_palette.R
\name{scale_color_palette}
\alias{scale_color_palette}
\title{Color Scale for Colorblind-Friendly Palette}
\usage{
scale_color_palette(index = NULL, ...)
}
\arguments{
\item{index}{A numeric vector specifying the indices of colors to use from the predefined palette. Values must be between 1 and 8. If NULL (the default), all colors are used.}

\item{...}{Additional arguments passed to \code{scale_color_manual}, such as \code{name} for legend title, or \code{labels} for legend labels.}
}
\value{
A ggplot2 color scale function with the colorblind-friendly palette.
}
\description{
This function applies a predefined colorblind-friendly palette to color aesthetics in \code{ggplot2} plots. You can select specific colors from the palette by specifying the \code{index} parameter.

The default palette includes the following colors:
1 - Blue, 2 - Vermilion, 3 - Bluish Green, 4 - Orange, 5 - Sky Blue, 6 - Yellow, 7 - Reddish Purple, 8 - Grey.
}
\details{
\code{scale_color_palette} provides a color scale that uses a colorblind-friendly palette, wrapping the \code{scale_color_manual} function.
}
\examples{
library(ggplot2)
bp <- ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
  geom_point(size = 5)
bp + scale_color_palette()

bp + scale_color_palette(index = c(4, 5, 7),
 labels = c("A", "B", "C"),
  name = "New Legend Title")

}
\references{
Based on \url{https://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette} and \url{https://paulgp.github.io/beamer_tips.pdf}.
}
