% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_palette2.R
\name{scale_fill_palette2}
\alias{scale_fill_palette2}
\title{wrapper for scale_fill_manual function}
\usage{
scale_fill_palette2(index = NULL, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{scale_fill_manual}. Mainly names and labels (see example).}

\item{col_order}{if col_order is NULL (the default), the function uses the color by the default order, col_order argument is used to override this behavior.}
}
\description{
This is a wrapper function for the scale_fill_manual function. It uses pre defined a colorblind-friendly palette
}
\details{
This function is heavily inspired by ggplot's \code{scale_fill_brewer} function. The palette includes the follwoing colors: 1 - Blue ; 2 - Red ; 3 - Black ; 4 - Orange ; 5 - Sky Blue ; 6 - Yellow ; 7 - Reddish Purple ; 8 - Teal Green ; 9 - Grey.
}
\examples{
scale_fill_palette()
library(ggplot2)
bp <- ggplot(iris, aes(Species, Sepal.Length)) +
geom_boxplot(aes(fill = Species)) +
theme_minimal() +
theme(legend.position = "top")
bp # Regular colors by ggplot2
bp + scale_fill_palette2()
bp + scale_fill_palette2(index = c(2,5,3),labels=c("A","B","C"),name = "New Legend Title")
}
\references{
Based on \url{https://davidmathlogic.com/colorblind/#\%230065A7-\%23961d24-\%23000000-\%23E69F00-\%2356B4E9-\%23CC79A7-\%23009988-\%23BBBBBB}, \url{https://personal.sron.nl/~pault/}, and \url{https://paulgp.github.io/beamer_tips.pdf}
}
