% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_as_date.R
\name{fast_as_date}
\alias{fast_as_date}
\title{Fast change to Date class}
\usage{
fast_as_date(x, ...)
}
\arguments{
\item{x}{a vector to be converted into date.}

\item{...}{Additional arguments passed to \code{as.Date.character}.}
}
\value{
A vector of Dates
}
\description{
This function coverts a character or factor vector into a vector of dates.
}
\examples{
fast_as_date(c("2020-10-15","2020-12-03"))
fast_as_date(c("10/15/20","12/03/20"),format='\%m/\%d/\%y')
}
\references{
Based on \url{https://towardsdatascience.com/tricks-in-r-to-boost-your-productivity-8c977242c69c}
}
