% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_palette.R
\name{scale_color_palette}
\alias{scale_color_palette}
\title{wrapper for scale_color_manual function}
\usage{
scale_color_palette(index = NULL, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{scale_color_manual}. Mainly names and labels (see example).}

\item{col_order}{if col_order is NULL (the default), the function uses the color by the default order, col_order argument is used to override this behavior.}
}
\description{
This is a wrapper function for the scale_color_manual function. It uses pre defined a colorblind-friendly palette
}
\details{
This function is heavily inspired by ggplot's \code{scale_fill_brewer} function. The palette includes the follwoing colors: 1 - Blue ; 2 - Vermilion ; 3 - Bluish Green ; 4 - Orange ; 5 - Sky Blue ; 6 - Yellow ; 7 - Reddish Purple ; 8 - Grey.
}
\examples{
library(ggplot2)
bp <- ggplot(iris, aes(x=Sepal.Width, y=Sepal.Length,color = Species)) +
geom_point(size=5) +
theme_minimal() +
theme(legend.position = "top")
bp # Regular colors by ggplot2
bp + scale_color_palette()
bp + scale_color_palette(index = c(4,5,7),labels=c("A","B","C"),name = "New Legend Title")
}
\references{
Based on \url{https://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette}, and \url{https://paulgp.github.io/beamer_tips.pdf}
}
